requirements.txt
ecdsa
mnemonic
pytest
cryptography
fastapi
httpx
rocketry
pyjwt[crypto]
pydantic-settings